+++ my-app/src/App.tsx
@@ -4,8 +4,8 @@ import Header from "./component/header/Header";
 import Clicker from "./component/clicker/Clicker";
 import Registration from "./component/registration/Registration";
 import Login from "./component/login/Login";
-import Posts from "./component/posts/Posts";
-import Post from "./component/postPage/PostPage";
+import PostsPage from "./component/postsPage/PostsPage";
+import PostPage from "./component/postPage/PostPage";
 
 import "./App.scss";
 
@@ -32,8 +32,8 @@ const App: React.FC = () => {
                             element={<Registration />}
                         />
                         <Route path="/posts">
-                            <Route index element={<Posts />} />
-                            <Route path=":id" element={<Post />} />
+                            <Route index element={<PostsPage />} />
+                            <Route path=":id" element={<PostPage />} />
                         </Route>
                         <Route path="*" element={<Navigate to={"/posts"} />} />
                     </Routes>
