+++ my-app/src/component/postsPage/PostsFilterReducer.ts
