+++ my-app/src/component/postsPage/PostsPage.scss
