+++ my-app/src/component/postsPage/card/PostCard.scss
