+++ my-app/src/component/postsPage/card/PostCard.tsx
